+++ builder.ts
@@ -6,7 +6,7 @@ import type {ConnectionRequirement} from './api_types';
 import type {Format} from './types';
 import type {Formula} from './api';
 import type {FormulaDefinitionV2} from './api';
-import type {MetadataFormula} from './api';
+import type {MetadataFormulaDef} from './api';
 import type {ObjectSchema} from './schema';
 import type {ParamDefs} from './api_types';
 import type {SyncFormulaDef} from './api';
@@ -14,9 +14,11 @@ import type {SyncTable} from './api';
 import type {SyncTableOptions} from './api';
 import type {SystemAuthentication} from './types';
 import type {SystemAuthenticationDef} from './types';
+import {isDynamicSyncTable} from './api';
 import {makeDynamicSyncTable} from './api';
 import {makeFormula} from './api';
 import {makeSyncTable} from './api';
+import {maybeRewriteConnectionForFormula} from './api';
 import {wrapMetadataFunction} from './api';
 
 /**
@@ -32,7 +34,7 @@ export function newPack(definition?: Partial<BasicPackDefinition>): PackDefiniti
   return new PackDefinitionBuilder(definition);
 }
 
-class PackDefinitionBuilder implements BasicPackDefinition {
+export class PackDefinitionBuilder implements BasicPackDefinition {
   formulas: Formula[];
   formats: Format[];
   syncTables: SyncTable[];
@@ -42,6 +44,8 @@ class PackDefinitionBuilder implements BasicPackDefinition {
   systemConnectionAuthentication?: SystemAuthentication;
 
   formulaNamespace?: string;
+
+  private _defaultConnectionRequirement: ConnectionRequirement | undefined;
 
   constructor(definition?: Partial<BasicPackDefinition>) {
     const {
@@ -63,7 +67,10 @@ class PackDefinitionBuilder implements BasicPackDefinition {
   }
 
   addFormula<ParamDefsT extends ParamDefs>(definition: FormulaDefinitionV2<ParamDefsT>): this {
-    const formula = makeFormula(definition);
+    const formula = makeFormula({
+      ...definition,
+      connectionRequirement: definition.connectionRequirement || this._defaultConnectionRequirement,
+    });
     this.formulas.push(formula as any); // WTF
     return this;
   }
@@ -76,7 +83,15 @@ class PackDefinitionBuilder implements BasicPackDefinition {
     connectionRequirement,
     dynamicOptions = {},
   }: SyncTableOptions<K, L, ParamDefsT, SchemaT>): this {
-    const syncTable = makeSyncTable({name, identityName, schema, formula, connectionRequirement, dynamicOptions});
+    const connectionRequirementToUse = connectionRequirement || this._defaultConnectionRequirement;
+    const syncTable = makeSyncTable({
+      name,
+      identityName,
+      schema,
+      formula,
+      connectionRequirement: connectionRequirementToUse,
+      dynamicOptions,
+    });
     this.syncTables.push(syncTable);
     return this;
   }
@@ -85,15 +100,18 @@ class PackDefinitionBuilder implements BasicPackDefinition {
   addDynamicSyncTable<ParamDefsT extends ParamDefs>(definition: {
     name: string;
     identityName: string;
-    getName: MetadataFormula;
-    getSchema: MetadataFormula;
+    getName: MetadataFormulaDef;
+    getSchema: MetadataFormulaDef;
     formula: SyncFormulaDef<ParamDefsT>;
-    getDisplayUrl: MetadataFormula;
-    listDynamicUrls?: MetadataFormula;
+    getDisplayUrl: MetadataFormulaDef;
+    listDynamicUrls?: MetadataFormulaDef;
     entityName?: string;
     connectionRequirement?: ConnectionRequirement;
   }): this {
-    const dynamicSyncTable = makeDynamicSyncTable(definition);
+    const dynamicSyncTable = makeDynamicSyncTable({
+      ...definition,
+      connectionRequirement: definition.connectionRequirement || this._defaultConnectionRequirement,
+    });
     this.syncTables.push(dynamicSyncTable);
     return this;
   }
@@ -133,6 +151,49 @@ class PackDefinitionBuilder implements BasicPackDefinition {
 
   addNetworkDomain(...domain: string[]): this {
     this.networkDomains.push(...domain);
+    return this;
+  }
+
+  setDefaultConnectionRequirement(connectionRequirement: ConnectionRequirement): this {
+    this._defaultConnectionRequirement = connectionRequirement;
+
+    // Rewrite any formulas or sync tables that were already defined, in case the maker sets the default
+    // after the fact.
+    this.formulas = this.formulas.map(formula => {
+      return formula.connectionRequirement ? formula : maybeRewriteConnectionForFormula(formula, connectionRequirement);
+    });
+    this.syncTables = this.syncTables.map(syncTable => {
+      if (syncTable.getter.connectionRequirement) {
+        return syncTable;
+      } else if (isDynamicSyncTable(syncTable)) {
+        return {
+          ...syncTable,
+          getter: maybeRewriteConnectionForFormula(syncTable.getter, connectionRequirement),
+          // These 4 are metadata formulas, so they use ConnectionRequirement.Required
+          // by default if you don't specify a connection requirement (a legacy behavior
+          // that is confusing and perhaps undesirable now that we have better builders).
+          // We don't know if the maker set Required explicitly or if was just the default,
+          // so we don't know if we should overwrite the connection requirement. For lack
+          // of a better option, we'll override it here regardless. This ensure that these
+          // dynamic sync table metadata formulas have the same connetion requirement as the
+          // sync table itself, which seems desirable basically 100% of the time and should
+          // always work, but it does give rise to confusing behavior that calling
+          // setDefaultConnectionRequirement() can wipe away an explicit connection
+          // requirement override set on one of these 4 metadata formulas.
+          getName: maybeRewriteConnectionForFormula(syncTable.getName, connectionRequirement),
+          getDisplayUrl: maybeRewriteConnectionForFormula(syncTable.getDisplayUrl, connectionRequirement),
+          getSchema: maybeRewriteConnectionForFormula(syncTable.getSchema, connectionRequirement),
+          listDynamicUrls: maybeRewriteConnectionForFormula(syncTable.listDynamicUrls, connectionRequirement),
+        };
+      } else {
+        return {
+          ...syncTable,
+          getter: maybeRewriteConnectionForFormula(syncTable.getter, connectionRequirement),
+          getSchema: maybeRewriteConnectionForFormula(syncTable.getSchema, connectionRequirement),
+        };
+      }
+    });
+
     return this;
   }
 }